feat: Updated UI
